requirements.txt
python-magic
pytest